Enabled and fixed a _lot_ of extra warnings and added the '-Werror' flag to handle every compile warning as an error

+++ include/cli_parser_base.hpp
@@ -54,7 +54,7 @@ class CLIParserBase {
     CLIParserBase(int argc, char *argv[]) {
         // Convert the char* argv[] to a vector of strings
         args.reserve(argc - 1);
-        for (size_t i = 1; i < argc; ++i) {
+        for (int i = 1; i < argc; ++i) {
             args.emplace_back(argv[i]);
         }
     }
